Merge less-work-to-just-port-it-7827-3: port styles to py3
Author: glyph

Reviewer: hawkowl, adiroiban

Fixes: #7827

twisted.persisted is not a good package. don't use it. But, some things
within Twisted still use it, and attempting to work around its absence
while porting those things to python 3 was taking more time than just
porting the code, so that's what this change does.  some parts of
it (such as crefutil) are actually used by reasonable things we intend
to keep maintaining, such as jelly, so this brings the quality of the
code up a little bit and improves coverage as well.

git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@45146 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb