Merge hostname-endpoint-dirty-reactor-8223: Clean up reactor after cancellation
Author: warner

Reviewer: glyph

Fixes: #8223

As noted in https://twistedmatrix.com/trac/ticket/8014#comment:15 , when a
HostnameEndpoint's Deferred is cancelled while waiting for a connection to
occur, it leaves the DelayedCall running. While this might not affect normal
applications, any trial tests that hit this case will fail, with a "Dirty
Reactor Error".

git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@46874 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb