**NOTE: THIS IS NOT WORKING 100%**
This patch introduces support for ECDH(E) and ECDSA, but I'm having some trouble getting ECDSA to work in all cases.

It works fine when using a Twisted SSH server and Conch, as well when using a OpenSSH server and Conch, but it only works about 40% of the time when using a Twisted SSH server and an OpenSSH client.

This part has me perplexed. It gives a "bad format" error, but not specifically where. I've tried looking for differences using Wireshark and various changes in the code, but have been unable to find out what is going wrong or where. To make matters even more confusing is the fact that it works about 40% of the time. I would expect it to work either 100% or 0% of the time.

So because I am at a loss of what could be happening, and that my time available to work on this is diminishing, I'm posting this hoping someone can finish it up and show me what's wrong.

Included in this patch is trial_out.txt which is the output of the trial tests.

It does fail on one instance: twisted.conch.test.test_endpoints.NewConnectionTests.test_mismatchedHostKey, where it was expected to fail, but didn't.
I haven't been able to figure out what causes this to fail, or even if it is a true failure or not.