Skip test_largeLong() on Python 3
On Python 3, long is gone, and int is an unlimited size, so
this test is meaningless there.
In a previous attempt to port this test to Python 3,
I tried to change sys.maxint to sys.maxsize, but
this broke the test running on Python 2 on Windows,
because on Windows sys.maxint != sys.maxsize.
Skipping this test on Python 3 is the better way to go.