There is no /dev/tty on Windows, disable new tests on it.
getPassword and getGenericAnswers have never worked on Windows, and the new
tests for those fails on Windows.
Bugs 6409 and 6410 are about changing the way these functions works, so
they are not blocking, and that will also solve the Windows problem.
Making an interim fix in the meantime seems a bit pointless.