address garbage collection issues
test failures on 3.3 are the result of some weird flaw in the garbage
collector which pollutes gc.garbage and causes later tests to fail; not
present on later versions of Python.

This change, however, breaks any errant circular references by
triggering connectionLost, and eliminates unnecessary Deferreds and
associated round-trip traffic in the streaming producer portions of PB
so that tests don't spuriously fail.