test_producerSSLFromStart should rely on handshakeCompleted.
Losing the connection from the client ProducerProtocol's
connectionMade allowed it to send an RST prior to the server calling
accept(2), because loseConnection before a handshake completes becomes
abortConnection.  On FreeBSD this resulted in t.i.tcp.Port.doRead
terminating early with ECONNABORTED and led the test to hang.

handshakeCompleted ensures that the loseConnection happens after the
server protocol's constructed, allowing it to fire its _done callback
and conclude the test.