Up arrow on a partial line appends bytes, not a list, to the history.
HistoricalRecvLine.handle_UP used to append the current line as a list
to the history, while HistoricalRecvLine.handle_RETURN appended a
string.  This was OK because HistoricalRecvLine._deliverBuffer
iterated over a slice that excluded the last character, and then
accessed the last character directory.  In Python 2 byte strings and a
list of individual characters both yield individual characters when
iterated over or indexed, but in Python 3 they yield integers.

1aac4c14 addressed iteration by using iterbytes, but assumed the
provided buf object was a byte string and attempted to extract a byte
string of length 1 via slice notation.  This eventually results in a
TypeError when buf is a list.

History items are immutable, so it's best to make them strings.
handle_UP now does that.

Thanks @exarkun for the diagnosis!