OpenSSL >= 1.1.0 fails with EOF when a client cannot complete a handshake.
Terminating a TLS connection after the client sends a ClientHello but
before it completes a handshake always fails.  The tests in this
commit terminate client connections in this way.

Earlier versions of OpenSSL fail the BIO flush with "SSL routines: ssl
handshake failure" error in ssl3_read_bytes.  The affected tests
expected this by only trapping Error.

The TLS state machine rewrite in 1.1.0 (see
https://www.openssl.org/news/openssl-1.1.0-notes.html) presumably
resulted in the BIO flush failing with EOF.