_ClientMachine: pass the Failure into _connectionFailed
This doesn't change behavior at all, but prepares for a later change which
does. Up until now, none of the responses to a failed connection really cared
about the actual failure object, but they will in the future. This patch
delivers the Failure to the (Automat) Input, but then changes all the Outputs
to ignore it.

Some of the Outputs must be replicated to accomodate the new method
signature. For example, _wait() takes no arguments, because it doesn't need
any. Before this change, when _connectionFailed provided no arguments, we
could have _connecting.upon(_connectionFailed) call a zero-argument _wait().
But now that _connectionFailed includes a Failure as an argument, we need to
provide a new _ignoreAndWait(f) that does the same thing as _wait() but knows
how to discard the argument. The original _wait() is still used by a
different pathway that doesn't involve a Failure object:
_connected.upon(_clientDisconnected)

(unfortunately we can't use default argument values to share the same method
between the pathway that adds a Failure and the one which doesn't: Automat is
very picky about the method signatures being completely identical)

In some cases, the body of the method was moved to a new shared helper
function, which can be called by both entry points. In others, where the body
only had one line to begin with, it was simply duplicated.