ClientService: add failOnFirstFailure=
If set to True, the first connection attempt that the ClientService makes is
special: a failure in this attempt will halt the reconnection logic, and will
deliver the ConnectionFailed error to any Deferreds obtained via
whenConnected.

This allows applications to react differently to initial connection problems,
such as a DNS error, or a complete lack of internet connectivity. If the
first attempt fails, additional attempts are unlikely to do any better. For
some applications, silently re-attempting the doomed connection forever may
cause more confusion than simply emitting an error and terminating
immediately.