working parse implemented for comma sep field-values
headers in general, when receiving byte strings that are comma separated, do not parse appropriately as separate field values until this point in the client. When item is b'1, 1' or b'1, 2', split and set header to list of field-values from field-value byte string. Side-Note: integer conversion is deferred until later, leaving uniqueness comparison for set operation to be done on field-value byte strings.

Signed-off-by: Evan.Mattiza <emattiza@gmail.com>