Add hostname caching policy to twisted.web.client
The HostnameCachingHTTPSPolicy is an IPolicyForHTTPS that caches
connection creators for reuse in subsequent requests.  The cache size is
determined by the cacheSize parameter and defaults to 20.  When a cache
miss occurs, an entry for the new host is created and added to the
cache.  The cache is then purged and the oldest entry removed.