Passes on Python 3: test_imap.IMAP4ServerTests.testLSub, testList
MemoryAccount now stores mailbox names as text; _parseMbox was lifted
out of IMAP4Server so that MemoryAccount can use it.

The current API allows users to create and query mailboxes as text in
Python 3 via the imap4-utf-7 codec.  At the same time, the results are
returned as native strings on Python 2, so that a LIST response has
the mailbox and delimiter as bytes.  However, on Python 3,
IMAP4Client.list and .lsub should return text to maintain symmetry
with the outbound APIs.  This is now achieved with version-specific
decoding.

References should also be text, because they refer to mailboxes, which
are text.