Passes on Python 3: twisted.mail.test.test_imap.IMAP4ServerTests.testNamespace
The current API allows users to create and query mailboxes as text in
Python 3 via the imap4-utf-7 codec.  At the same time, results are
returned as native strings on Python 2, so that the namespace query
response has mailbox namespaces and delimiters as bytes.  However, on
Python 3, IMAP4Client.namespace return text to maintain symmetry with
the outbound APIs.  This is now achieved with version-specific
decoding.