Passes on Python 3: test_imap.NewFetchTests.testFetchEnvelope
Note that IMAP4Server expects IMessage.getHeaders to return a
dictionary composed of native strings now.  There's a practical reason
and a conceptual reason for this.

The practical reason is that Python 3's email packages work in terms
of native strings, and most implementers of IMessage probably use
these.

The theoretical reason is that RFC 6532 allows UTF-8 in email headers,
and RFC 6855 extends IMAP4rev1 to support this.  imap4 can and should
support Unicode strings beyond those that are ASCII congruent Given an
RFC 6855 implementation.

IMAP4Client returns envelope data as native strings decoded with
charmap on Python 3.  This will cause mojibake, but maintains the
native-string API.