twisted.mail.imap4.MessageSet sorts * larger than any message.
RFC 3501 declares that * is not necessarily the largest UID, just the
most recent.  However treating * as the largest UID is consistent with
Cyrus' interpretation of *, and means None can be replaced with
float('inf') for all comparisons.

None now ends up in a consistent location in MessageSet.ranges, and is
never compared to an integer.  This makes the implementation more
robust and easier to port to Python 3, where numeric comparisons to
None fail with a TypeError.