Capture the signal that caused the reactor to exit.
Currently, if a signal causes the reactor to exit, that signal is lost
to the user.  This poses a problem for clients interested in the signal
to ensure a proper exit code.

Capturing the signal in a class attribute allows interested parties
(e.g. twistd) to obtain the value for later use.