Select reactor in startReactor
If the caller hasn't selected a reactor, choose the default now so that
the ApplicationRunner has access to the reactor after
runReactorWithLogging returns.  Without a reference to the reactor, the
application runner has no way of accessing the exitSignal attribute to
properly set the exit status.

The redundancy here (runReactorWithLogging also sets the reactor, so
that could have been lifted) to avoid modifying runReactorWithLogging.
While twisted doesn't currently use it in many places, it is a public
api and may be used by existing client code.