Make _exitSignal class attribute private on ReactorBase.
A concern raised during review is that this attribute exists on some
reactors, but not others, but isn't documented in any interface.  This
characteristic is true of signal handling by reactor implementations --
only some inherit _SignalReactorMixin and handle signals.  Since this
attribute is coupled to the same functionality, its visibility should
match that of signal handling.