Handle exit within Twisted.run.
If when Runner.run returns, determine whether a signal caused the
reactor to stop and deliver appropriately.  Moving this logic into run
allows for better testing (less patching) without trading anything in
return.

In testing, patch as little as possible, specifically, the
_exitWithSignal and startLogging methods must be patched.  The latter
because even delivering a dummy factory will result in a new observer
that will cause other, unrelated tests to fail.