ReactorBase implements _ISupportsExitSignalCapturing
By implementing this interface, ReactorBase commits to setting
_exitSignal with the value of any captured exit signal.

There is an alternative way to implement this, in the reactor subclasses
themselves, so it's worth documenting why I chose to implement the
interface in ReactorBase.  A brief rationale follows.

The signals referenced in this commit are exit signals, i.e. those
commonly used to exit a program such as SIGTERM, SIGINT, and SIGBREAK.
Specifically, this does not refer to SIGCHILD.

Currently, two primary actors are involved in signal handling by
reactors -- _SignalReactorMixin and ReactorBase.  The current pattern
is to inherit the mixin at the specific reactor (e.g. PosixReactorBase),
relying on the fact that _SignalReactorMixin is designed to work with a
subclass of ReactorBase.  Specifically, it expects the sigInt, sigTerm,
and sigBreak methods to exist (these methods are not part of any
interface).

This means that while the decision to install the handlers is made by
inheriting _SignalReactorMixin, the implementation of the handler lives
in ReactorBase.  Thus, rather than require each individual reactor to
implement their own (possibly delegating) sigInt, sigTerm, etc.,
implementing _ISupportsExitSignalCapturing in ReactorBase keeps the
responsibility for capturing the integer signal with the code that
handles the signal.