Wrap an IPolicyForHTTPS instead of duplicating behavior.
Rather than duplicate the existing behavior of
BrowserLikePolicyForHTTPS, allow clients to pass in any arbitrary
IPolicyForHTTPS and cache the returned connection creator.  This lets
the various existing (and future) policies manage the creation of the
connector in whatever way suits there needs, but allows for reuse of the
connector for the same hostname.