Add test case to handle an exception thrown in WSGIRequest handling.
This was caught during review.  There was logging in the test that had
never been covered because it is extremely unlikely that any existing
code path would lead to it.  However, the combination of a poorly
implemented third party reactor and a WSGI application that throws an
exception could lead to the exception making its way back to the
threadpool.