Set Request.client and Request.host in Request.__init__.
Previously, Request.getClientIP raised an AttributeError when it was
called before requestReceived where these were set.  They're retrieved
from the channel that's passed to Request.__init__, so setting them
there eliminates the possibility of an AttributeError.

As a result, it's no longer necessary to call requestReceived before
getClient in tests.