t.w.h.Request.getRequestHostname handles UNIX domain sockets.
A Request without a Host header received over a UNIX domain socket
cannot have a hostname.  Previously, such a request's
getRequestHostname method would fail with an AttributeError while
attempting to access the host attribute of the address returned by
getHost.

This commit introduces a new public exception, UnsupportedTransport,
that is instead raised when a Request's hostname cannot be determined.