Clarify _FileDescriptorReservation implementation and API.
Calling an MethodicalMachine input method in an output depends on
undefined behavior.  This commit replaces the public `acquire` input
with a new `reserve` method that uses private inputs to reserve a file
descriptor and transition back into the _unreserved state if that
fails with EMFILE.  The pure `available` transition was removed in
favor of a getter method to minimize the size of the state machine.