Refactor unformattable tests.
The review comment requested removal of the Unformattable() object in
the event key as it was extraneous to the test.  This is true, but it
does represent a valid test path on its own -- the path where the
object throws an exception when __repr__ is invoked.  Rather than
discard this case, I split the test into two tests.