Don't allow EPERM to loop accept(2) infinitely and test catch-all.
`accept(2)` is documented as failing with EPERM when "firewall rules
forbid connection".  That sounds like an error that's fatal for the
listening port.

But since 76b194f5b9e70883a7247d21102a746ffab21a82, Twisted has
retried `accept(2)` after EPERM failures.  Previously the `_accept`
refactoring assumed that an `accept(2)` that failed with EPERM would
eventually succeed, but there's no documentation, mailing list posts,
or other lore that confirms this.

Now instead of looping infinitely on `accept(2)`, the `_accept`
function receives an iterable that it iterates over to limit the
number of consecutive `accept(2)` calls it makes.  This also enables a
test to be written for its EPERM handling.

Also, test that a `socket.error` whose errno is not handled by
`_accept` is logged.