test_closePeerOnEMFILE allows ConnectionLost on client's end.
Whether the client's connectionLost method is called with
ConnectionDone or ConnectionLost depends on its internal state, which
the test cannot directly control; for example, a TLS handshake might
not have completed by the time the server disconnects the client.

Make the test more robust by allowing both ConnectionDone and
ConnectionLost.