Add a new formatEventWithTraceback API.
Currently, formatEvent does not handle tracebacks, yet it is commonly
used to format log events.  A common approach to dealing with this
shortcoming is to append a traceback to the result of formatEvent.
This is the approach taken by formatEventAsClassicLogText, and likely
some third party code as well.  Modifying the behavior of formatEvent
will cause duplicate tracebacks, which, while not necessarily the
worst error, will be confusing and is probably best avoided.

The alternate approach is a new API that allows code that calls
formatEvent directly to migrate deliberately to the new API.