Introduce a new eventAsText API.
eventAsText facilitates formatting a log event as unicode and
optionally including traceback, system, and timestamp information.
This API should be preferred over the existing formatEvent and
formatEventAsClassicLogText APIs.

One of the problems with the existing APIs is they do not always
include tracebacks, or handle multibyte/non-ascii encodings in the
tracebacks properly.  This API does both.

Update the narrative documentation to recommend this API over the
existing formatEvent API