Remove newline from eventAsText result.
This was raised during review.  There's no value in eventAsText
appending a newline to the resultant text.  The original goal was to
preserve the ending of the existing functions whose internals have
been replaced, but it isn't necessary for it to exist within
eventAsText to achieve that goal.

This change moves the newline out of eventAsText and back into the
appropriate functions (e.g. formatEventAsClassicLogText.