Keep format string static.
Rather than forcing the format string to be rebuilt on each call, have
it static and ensure that the appropriate whitespace/symbols are added
when the components (e.g. system) are set based on the include flags.

Raised during review as this has been hot path code in the past.