Merge jeremycline:9668-jeremycline-stdlog-findCaller-38-compat: Add the stackLevel kwarg to STDLibLogObserver._findCaller
Author: jeremycline
Reviewer: hawkowl
Fixes: ticket:9668

Python 3.8 adds a new keyword argument, stacklevel, to its findCaller
function in commit dde9fdbe4539 ("bpo-33165: Added stacklevel parameter
to logging APIs. (GH-7424)"). As Twisted is replacing this method with
its own method, logging with STDLibLogObserver fails in Python 3.8. This
patch adds the argument, but does not use it as there is already a
stackDepth instance variable and the stackInfo on the method is also
ignored.

This patch fixes the logger tests that currently fail in Python 3.8.

Signed-off-by: Jeremy Cline <jcline@redhat.com>