Support encoding to new OpenSSH private key format
This is the other half of 4b6a2a2d85bc1493d8d0d1eb90a3896ac220da5c, and
is a prerequisite for adding support for Ed25519 keys which OpenSSH only
ever writes in the new format.

OpenSSH refuses to load keys in this format that lack a trailing
newline, so I adjusted the test data accordingly.