Normalize Unicode passphrases for SSH keys
This is required by NIST 800-63B.  We choose NFKC normalization here.

It is possible, though unlikely, that some existing key files will have
passphrases that are the UTF-8 encoding of a Unicode string not in NFKC;
twisted.conch.ssh.keys.Key.fromString will fail to load those keys if
given such a passphrase as a Unicode string.  In that case the
workaround is to provide the passphrase as bytes instead.