Run pycodestyle-twisted against merge-base with trunk
`git diff trunk` includes changes on trunk since the common ancestor of
the current branch and trunk, which is unreasonably confusing if you're
looking at `tox -e lint` output on an out-of-date branch.  Instead, use
`git diff trunk...`, which diffs against the common ancestor of the
current branch and trunk.