fix exception from multiple calls to FileBodyProducer.stop
The current specification for IBodyProducer states that
implementations need to be written to allow simultaneous calls to
pauseProducing, resumeProducing, and stopProducing.  FileBodyProducer
currently only complies with that for pause and resumeProducing (where
comply is interpreted here as not raising an error).

Instead of allowing the TaskFinished exception to be raised on
multiple calls to stopProducing, catch and drop the exception.  The
use of "pass" is not ideal, but this avoids an internal book-keeping
variable to track the stop state and a branch for the normal flow.
The use of "pass" to handle stopped also appears within the inner
maybeStopped function of startProducing, so it isn't without precedent.

This has been discussed in several forums:
* https://github.com/twisted/treq/issues/185#issuecomment-555596421
* https://twistedmatrix.com/pipermail/twisted-python/2017-November/031730.html

Related ticket(s): 7457

Regression risk: Existing uses of FileBodyProducer that rely on
multiple calls to stopProducing to raise an exception will break.