Merge 9652-wiml-mktime: Allow mktime() to raise EOVERFLOW if isdst=1 and there's no DST.
Author: wiml
Reviewer: hawkowl
Fixes: ticket:9652

The mktime() call in versions of glibc starting with 2.28 will return
EOVERFLOW if asked to interpret a tuple whose isdst field is 1 in
timezones that don't observe Daylight Savings Time. We now accept that
mktime() *either* ignores isdst in those timezones *or* raises
EOVERFLOW, both of which appear to be legal behavior.