xmlrpc: verify HTTPS certificates by default
Use ssl.optionsForClientTLS() when opening the HTTPS connection with
connectSSL(). With this change, Twisted will use ssl.platformTrust() to
verify the SSL connection.

Add a new trustRoot argument to the Proxy class constructor. This allows
users to control how to trust the HTTPS certificate.

Raise ssl.SSL.Error in QueryProtocol instead of silently skipping it.