Simplify fetching of fancybasename in FancyStrMixin
Instead of checking for the existence of the attribute, we get it
with a default value. This statisfies mypy, but it's also easier
to understand for the human reader, in my opinion.

There is a subtle difference in functionality: if fancybasename
exists but is the empty string, the old could would fall back
on the class name, while the new code uses the empty string.

I don't think this will cause problems:
- none of the fancybasename values in Twisted's source are empty
- the new code matches the existing documentation