Disable pycodestyle for now
The incremental code style checking is blocking automated substitutions
on the source tree, since any change close to a piece of pre-existing
code that is not style compliant will make pycodestyle complain. This
happens so many times that correcting these manually is too much work.

After adopting Black, there will either be no need for style checking
or it can be done with a much smaller ruleset.