Deprecate compat.bytesEnviron() and remove its uses
There were two callers inside Twisted: test_compat and test_process.

The code in test_compat is the unit test and it checks that the
returned environment dictionary indeed contains only keys and values
of type 'bytes'. I left this as-is.

The code in test_process was weird: this is a Windows-specific test,
while bytesEnviron() claims to be POSIX-only in its docstring. It
turns out that bytesEnviron() doesn't break when executed on
Windows, but it produces an environment with keys and values as
strings instead of bytes.

The test passes despite this mismatch, but to reduce the confusion
I removed the call and replaced it with a simple copy of os.environ,
so it is more clear what is actually being tested here: an environment
with mostly string entries and one entry in bytes.