Remove override of assertRegex()
Since the override now does nothing except calling the superclass
implementation, there is no need to keep it.

It could be argued that the docstring has value, but I think it's
more consistent to refer to the standard library documentation
instead, since that will be kept up-to-date. Also it is more
consistent since for example assertNotRegex() exists in the
standard library but is not overridden here.