Fix TwistedModuleVisitor
When pydoctor switched from using the `compiler` module to `ast`,
the names of the visitor methods changed.

Recent pydoctor versions ignore inner functions, so we can't assume
that the current node will always be added to `contents`.

Also, don't bother making a copy of the decorators, since the
getDeprecated() function does not modify the list.