Fix broken links in documentation
The links to IReactorSocket.adoptDatagramPort in the UDP howto had
a space inside the link target that shouldn't be there.

A link in the compatibility policy was missing the closing '>'
around the label.

Two links in the names client tour were missing a space between
the link target and label. While Sphinx will parse it correctly
without the space, it is an unnecessary inconsistency that makes
it harder to convert the links automatically.