Fix API links that attempt to link to inherited targets
Intersphinx does not have a notion of inheritance, so the link must
point to the actual definition.

In some cases the actual definition has a private name, so maybe
pydoctor should create a public alias for those. But for now,
a working link to a private name is better than a broken link.