Explicitly define `__all__` value in deprecated proto_helpers module
Until now the `__all__` of `twisted.internet.testing` was cloned, but
that could also contain names that were added after the deprecation
and which are not exported by the deprecated module.

Also pydoctor can only handle literal values assigned to `__all__`
and will warn about other expressions in future releases.

While it is a bit odd to have private names in `__all__`, they are
preserved to avoid any backwards compatibility problems.