Wrap runtime updates of TwistOptions docstrings in a function
These docstrings are primarily command line help for the `Options`
mechanism. By wrapping the update in a function, we tell pydoctor
that it doesn't need to warn about being unable to analyze the
docstring's actual value.

The warning existed in pydoctor for a while, but in the next release
it will become enabled at the log level that Twisted uses.