Sessions: use the site reactor, not the global one
Expose the reactor associated with the `Site` by renaming the `_reactor`
attribute to `reactor` in its superclass `HTTPFactory`.

Make the `Session` initializer take the reactor from its *site* argument
when the explicit *reactor* argument isn't passed. It used to use the
global reactor.

There were a few tests that called the `Session` constructor with
incorrect arguments. It looks like the `uid` used to be an integer, but
it's bytes now. Also a few had the *uid* and *site* arguments swapped.

I also added some missing documentation.