Fix renderElements()'s errback
The flatten() function returns a `Deferred`, so we can't call
`encode()` on it. We wouldn't need to anyway, since the flattened
data is already `bytes` (an UTF-8 encoded string).

This got broken in the Python 3 port, commit 55f72311 to be exact.
Since the failure was swallowed by the finish handler after all
output had already been produced, no-one noticed.