Improve type annotations for DeferredList and gatherResults()
Since `DeferredList` creates a new list from the given `Deferred`s,
there is no need to require an actual `list` to be passed: any
iterable is fine.

The previous annotations used `object` as the result type, but as
the type argument is contravariant, `Deferred[object]` was the only
valid specialization that could be passed.

With the new annotations, mypy can derive the result type from
the type of the argument passed to the `DeferredList` constructor
or to `gatherResults()`'s first argument.

A complication was the `fireOnOneCallback` flag, which selects
between two different result types. I don't think it is possible
to express this for a class in Python's type system, so I added
a workaround using overloaded function definitions.