Make renderElement() accept any IRequest
The only reason the request was limited to `twisted.web.server.Request`
was that the error handler used the `site` attribute to check whether
it should output a detailed backtrace or a minimal error message.

If a different `IRequest` implementer is passed, we pick the safe
option and output a minimal error message.